# platform/contracts/openapi/v1/wego-api.yaml
openapi: 3.1.1
info:
  title: Wego API
  version: 1.0.0
  description: >-
    Version 1 of the product-neutral Wego HTTP contract. WEGO-001 adds the
    first authenticated operations (login, logout, current identity);
    WEGO-002 adds the first business-domain operations (diving offerings and
    staff-created bookings).
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
    description: Current client instance
...
paths:
...
  /api/v1/tours-operator/tours/by-slug/content:
    get:
      operationId: getToursOperatorPublicTourContent
      summary: Published tour content for the public site
      description: >
        PUBLISHED text in the requested locale (else published English, with
        servedLocale reporting which), published facts, stops that have both
        published coordinates and text, and rights-APPROVED media only.
      tags:
        - ToursOperator
      security: []
      parameters:
        - name: slug
          in: query
          required: true
          schema:
            type: string
        - name: locale
          in: query
          schema:
            $ref: "#/components/schemas/ToursOperatorContentLocale"
      responses:
        "200":
          description: Public content (fields are empty when nothing is published).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorPublicTourContent"
        "400":
          description: Unsupported locale.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
        "404":
          description: Tour not found or not active.
  /api/v1/tours-operator/staff/tours/{id}/content:
    get:
      operationId: getToursOperatorStaffTourContent
      summary: Draft and published content, facts and media for staff
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Staff content view.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorStaffTourContent"
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
  /api/v1/tours-operator/staff/tours/{id}/content/{locale}:
    put:
      operationId: saveToursOperatorTourContentDraft
      summary: Save the DRAFT content of one locale
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: locale
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ToursOperatorContentLocale"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToursOperatorTourContentDocument"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
  /api/v1/tours-operator/staff/tours/{id}/content/{locale}/publish:
    post:
      operationId: publishToursOperatorTourContent
      summary: Publish the DRAFT of one locale
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: locale
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ToursOperatorContentLocale"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToursOperatorPublishRequest"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
        "409":
          description: nothing_to_publish (no draft / nothing published) or draft_changed (the reviewed revision is outdated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
  /api/v1/tours-operator/staff/tours/{id}/content/{locale}/unpublish:
    post:
      operationId: unpublishToursOperatorTourContent
      summary: Remove one locale from the public site (draft kept)
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: locale
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/ToursOperatorContentLocale"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
        "409":
          description: nothing_to_publish (no draft / nothing published) or draft_changed (the reviewed revision is outdated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
  /api/v1/tours-operator/staff/tours/{id}/facts:
    put:
      operationId: saveToursOperatorTourFactsDraft
      summary: Save the DRAFT facts and stops
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToursOperatorTourFactsDocument"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
  /api/v1/tours-operator/staff/tours/{id}/facts/publish:
    post:
      operationId: publishToursOperatorTourFacts
      summary: Publish the DRAFT facts
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToursOperatorPublishRequest"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
        "409":
          description: nothing_to_publish (no draft / nothing published) or draft_changed (the reviewed revision is outdated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
  /api/v1/tours-operator/staff/tours/{id}/facts/unpublish:
    post:
      operationId: unpublishToursOperatorTourFacts
      summary: Remove facts from the public site (draft kept)
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
        "409":
          description: nothing_to_publish (no draft / nothing published) or draft_changed (the reviewed revision is outdated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
  /api/v1/tours-operator/staff/tours/{id}/media:
    put:
      operationId: replaceToursOperatorTourMedia
      summary: Replace the media list (approval kept only for unchanged id + path)
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: revision
          in: query
          description: Current staff mediaRevision. Required when the old or submitted list contains managed assets, including unlinking all images.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              maxItems: 30
              items:
                $ref: "#/components/schemas/ToursOperatorTourMediaInput"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour not found.
        "409":
          description: draft_changed; the whole media list changed after review.
  /api/v1/tours-operator/staff/tours/{id}/media/{mediaId}/approve:
    post:
      operationId: approveToursOperatorTourMedia
      summary: Approve usage rights of one media item
      tags:
        - ToursOperator
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: mediaId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToursOperatorPublishRequest"
      responses:
        "204":
          description: Done.
        "400":
          description: Invalid request (content rule, locale or media rule) or malformed body.
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/UnauthenticatedResponse"
        "403":
          $ref: "#/components/responses/ForbiddenResponse"
        "404":
          description: Tour or media item not found.
        "409":
          description: draft_changed — the media changed after review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToursOperatorContentErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Opaque session token issued by POST /api/v1/identity/login.
...
  schemas:
...
    ToursOperatorContentErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: string
          enum:
            - invalid_content
            - unsupported_locale
            - too_many_media
            - multiple_covers
            - duplicate_media_path
            - duplicate_media_id
            - english_alt_required
            - nothing_to_publish
            - draft_changed
            - unknown_managed_asset
            - managed_asset_mismatch
            - media_revision_required
        detail:
          type:
            - string
            - "null"
          description: The broken rule for invalid_content (field-level, never stored data).
...
    ToursOperatorContentLocale:
      type: string
      enum:
        - en
        - ar
        - ru
        - it
...
    ToursOperatorPublicTourContent:
      type: object
      additionalProperties: false
      required:
        - tourId
        - slug
        - requestedLocale
        - includes
        - excludes
        - knowBeforeYouGo
        - stops
        - media
      properties:
        tourId:
          type: string
          format: uuid
        slug:
          type: string
        requestedLocale:
          $ref: "#/components/schemas/ToursOperatorContentLocale"
        servedLocale:
          oneOf:
            - $ref: "#/components/schemas/ToursOperatorContentLocale"
            - type: "null"
        name:
          type:
            - string
            - "null"
        shortDescription:
          type:
            - string
            - "null"
        description:
          type:
            - string
            - "null"
        includes:
          type: array
          items:
            type: string
        excludes:
          type: array
          items:
            type: string
        knowBeforeYouGo:
          type: array
          items:
            type: string
        meetingPoint:
          type:
            - string
            - "null"
        facts:
          oneOf:
            - type: object
              additionalProperties: false
              required:
                - guideLanguages
              properties:
                childrenAllowed:
                  type:
                    - boolean
                    - "null"
                minimumAge:
                  type:
                    - integer
                    - "null"
                guideLanguages:
                  type: array
                  items:
                    type: string
                hotelPickup:
                  oneOf:
                    - $ref: "#/components/schemas/ToursOperatorHotelPickup"
                    - type: "null"
            - type: "null"
        stops:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - key
              - kind
              - latitude
              - longitude
              - name
            properties:
              key:
                type: string
              kind:
                type: string
                enum:
                  - STOP
                  - MEETING_POINT
              latitude:
                type: number
              longitude:
                type: number
              name:
                type: string
              description:
                type:
                  - string
                  - "null"
        media:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - path
              - width
              - height
              - isCover
              - alt
            properties:
              path:
                type: string
              width:
                type: integer
              height:
                type: integer
              isCover:
                type: boolean
              alt:
                type: string
...
    ValidationErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
        - message
      properties:
        error:
          type: string
          enum:
            - validation_failed
        message:
          type: string
          description: A safe, human-readable description of which rule failed.
...